+++ editor-app/src/app/antlr-editor/extensions/syntax-highlight/theme/dark.ts
