+++ editor-app/src/wasm/analyzer.ts
